+++ src/execute/engine.ts
@@ -36,11 +36,9 @@ import { pandocBuiltInFormats } from "../core/pandoc/pandoc-formats.ts";
 import { gitignoreEntries } from "../project/project-gitignore.ts";
 import { ensureFileInformationCache } from "../project/project-shared.ts";
 import { engineProjectContext } from "../project/engine-project-context.ts";
-import { Command } from "cliffy/command/mod.ts";
 import { getQuartoAPI } from "../core/api/index.ts";
 import { satisfies } from "semver/mod.ts";
 import { quartoConfig } from "../core/quarto.ts";
-import { initializeProjectContextAndEngines } from "../command/command-utils.ts";
 
 const kEngines: Map<string, ExecutionEngineDiscovery> = new Map();
 
@@ -385,42 +383,3 @@ export function projectIgnoreGlobs(dir: string) {
     gitignoreEntries(dir).map((ignore) => `**/${ignore}**`),
   );
 }
-
-export const engineCommand = new Command()
-  .name("engine")
-  .description(
-    `Access functionality specific to quarto's different rendering engines.`,
-  )
-  .stopEarly()
-  .arguments("<engine-name:string> [args...:string]")
-  .action(async (options, engineName: string, ...args: string[]) => {
-    // Initialize project context and register external engines
-    await initializeProjectContextAndEngines();
-
-    // Get the engine (now includes external ones)
-    const engine = executionEngine(engineName);
-    if (!engine) {
-      console.error(`Unknown engine: ${engineName}`);
-      console.error(
-        `Available engines: ${
-          executionEngines().map((e) => e.name).join(", ")
-        }`,
-      );
-      Deno.exit(1);
-    }
-
-    if (!engine.populateCommand) {
-      console.error(`Engine ${engineName} does not support subcommands`);
-      Deno.exit(1);
-    }
-
-    // Create temporary command and let engine populate it
-    const engineSubcommand = new Command()
-      .description(
-        `Access functionality specific to the ${engineName} rendering engine.`,
-      );
-    engine.populateCommand(engineSubcommand);
-
-    // Recursively parse remaining arguments
-    await engineSubcommand.parse(args);
-  });
